[fix] regression testを修正

+++ sources/arm_asm/05_asm/regression_test.py
@@ -6,21 +6,29 @@ import os
 
 filenames = ["test"]
 
+compile_utils = "gcc -c cl_utils.c".split()
+compile_main = "gcc -c main.c".split()
+link = "gcc cl_utils.o main.o -o test".split()
+
+
 def main():
+    # コンパイル
+    subprocess.run(compile_utils, shell=True)
+    subprocess.run(compile_main, shell=True)
+    subprocess.run(link, shell=True)
+
+    is_success = True
     for file_name in filenames:
-        compile_utils = "gcc -c cl_utils.c".split()
-        compile_main = "gcc -c main.c".split()
-        link = "gcc cl_utils.o main.o -o test".split()
         exec = f"test.exe ./test_input/{file_name}.ks".split()
 
-        # コンパイル、実行
-        subprocess.run(compile_utils, shell=True)
-        subprocess.run(compile_main, shell=True)
-        subprocess.run(link, shell=True)
+        # 実行
         subprocess.run(exec, shell=True)
 
         # ファイル読み込み、バイナリ比較
         with open(f"./test.bin", "rb") as ac_f, open(f"./test_expect/{file_name}.bin", "rb") as ex_f:
+
+            line_num = 0
+
             while True:
                 actual = ac_f.read(4).hex()
                 expect = ex_f.read(4).hex()
@@ -30,25 +38,35 @@ def main():
                 if actual != expect:
                     print("error")
                     print(f"File Name: {file_name}")
+                    print(f"line_num: 0x{line_num:08X}")
 
                     # 16進数ダンプ
-                    ac_iter = iter(list(actual.upper()))
-                    print("actual: ", end="")
-                    for a, b in zip(ac_iter, ac_iter):
-                        print(f"{a}{b} ", end="")
-                    print("")
-                    ex_iter = iter(list(expect.upper()))
-                    print("expect: ", end="")
-                    for a, b in zip(ex_iter, ex_iter):
-                        print(f"{a}{b} ", end="")
 
+                    # actual
+                    act_str = actual.upper()
+                    print("actual: ", end="")
+                    for one in [act_str[i:(i+2)] for i in range(4)]:
+                        print(one, " ", end="")
+                    print("")
+
+                    # expect
+                    exp_str = expect.upper()
+                    print("expect: ", end="")
+                    for one in [exp_str[i:(i+2)] for i in range(4)]:
+                        print(one, " ", end="")
+
+                    is_success = False
                     break
+
+                line_num += 0x4
 
     # ファイル削除
     os.remove("./main.o")
     os.remove("./cl_utils.o")
     os.remove("./test.exe")
-    os.remove("./test.bin")
+
+    if is_success:
+        os.remove("./test.bin")
